% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.regexr.character}
\alias{as.regexr.character}
\title{Coerce character to regexr}
\usage{
\method{as.regexr}{character}(x, names = TRUE, comments = TRUE,
  comments.below = TRUE, ...)
}
\arguments{
\item{x}{The \code{character} object.}

\item{names}{logical.  Should names be included in the \code{construct}
script?}

\item{comments}{logical.  Should comments be included in the \code{construct}
script?}

\item{comments.below}{logical.  Should comments be included below the regex
in the \code{construct} script?  If \code{FALSE} comments are placed behind
the regular expression chunks.}

\item{\ldots}{Ignored.}
}
\description{
Convert a regular expression to a commented \code{regexr} object.
}
\details{
character Method for as.regexr
}

