% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{comments<-.regexr}
\alias{comments<-.regexr}
\title{Set Comments For regexr}
\usage{
\method{comments}{regexr}(x) <- value
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{value}{The comment(s) to assign.}

\item{\ldots}{Ignored.}
}
\description{
\code{regexr} - Set comments for \code{regexr} object.
}
\details{
regexr Method for comments<-
}

