% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comments}
\alias{comments}
\alias{comments<-}
\title{Get/Set Comments From a regexr Object}
\usage{
comments(x, ...)

comments(x) <- value
}
\arguments{
\item{x}{A regexr object.}

\item{value}{The comment(s) to assign.}

\item{\ldots}{Ignored.}
}
\value{
\code{comments} - Returns a list of comments.
}
\description{
\code{comments} - Get the \code{comments} from a \code{regexr} object.

\code{comments<-} - Set the \code{comments} of a \code{regexr} object.
}
\examples{
minimal <- construct("a", "b", "c" \%:)\% "Comment #3")
minimal
comments(minimal)
comments(minimal)[2] <- "A comment"
comments(minimal)
}

