% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.regexr.R
\name{get_construct.reverse_construct}
\alias{get_construct.reverse_construct}
\title{Extract Script from \code{reverse_construct} to \code{construct} a
\code{regexr} Object.}
\usage{
\method{get_construct}{reverse_construct}(x, file = "", ...)
}
\arguments{
\item{x}{A \code{reverse_construct} object.}

\item{file}{A connection, or a character string naming the file to print to.
If "" (the default), \code{\link[base]{cat}} prints to the console unless
redirected by \code{\link[base]{sink}}.  Windows users may use
\code{file = "clipboard"} to copy the content to the clipboard. To print
as \code{character} use \code{file = NULL}.}

\item{\ldots}{Other arguments passed to \code{\link[regexr]{print.reverse_construct}}.}
}
\value{
Returns an auto-commented script used to \code{construct} a
\code{regexr} object.
}
\description{
Pulls the \code{reverse_construct} attribute from a \code{reverse_construct}.
This script can be assigned to an object and run in the console to create a
comments, named \code{regexr} object.
}
\details{
reverse_construct Method for get_construct
}

