% Generated by roxygen2 (4.0.2): do not edit by hand
\name{names<-.regexr}
\alias{names<-.regexr}
\title{Set Names of a regexr Regex Chunks}
\usage{
\method{names}{regexr}(x) <- value
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{value}{The comment(s) to assign.}

\item{\ldots}{Ignored.}
}
\description{
Set names of a \code{regexr} object's regular expression chunks.
}

