% Generated by roxygen2 (4.0.2): do not edit by hand
\name{names.regexr}
\alias{names.regexr}
\title{Get Names of a regexr Object}
\usage{
\method{names}{regexr}(x, ...)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\description{
Get names of a \code{regexr} object.
}

