% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.regexr.R
\name{print.reverse_construct}
\alias{print.reverse_construct}
\title{Prints a reverse_construct object}
\usage{
\method{print}{reverse_construct}(x, file = "", ...)
}
\arguments{
\item{x}{A \code{reverse_construct} object.}

\item{file}{A connection, or a character string naming the file to print to.
If "" (the default), \code{\link[base]{cat}} prints to the console unless
redirected by \code{\link[base]{sink}}.  Windows users may use
\code{file = "clipboard"} to copy the content to the clipboard.}

\item{\ldots}{Other arguments passed to \code{\link[base]{cat}}.}
}
\description{
Prints a reverse_construct object.
}

