% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regex-class.R
\name{print.subcom}
\alias{print.subcom}
\title{Prints a subcom object}
\usage{
\method{print}{subcom}(x, ...)
}
\arguments{
\item{x}{The subcom object.}

\item{\ldots}{Ignored.}
}
\description{
Prints a subcom object.
}

