% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{print.summary_regexr}
\alias{print.summary_regexr}
\title{Prints a summary_regexr object}
\usage{
\method{print}{summary_regexr}(x, ...)
}
\arguments{
\item{x}{The summary_regexr object.}

\item{\ldots}{Ignored.}
}
\description{
Prints a summary_regexr object.
}

