% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{print.unglued}
\alias{print.unglued}
\title{Prints an unglued object}
\usage{
\method{print}{unglued}(x, ...)
}
\arguments{
\item{x}{The unglued object.}

\item{\ldots}{Ignored.}
}
\description{
Prints an unglued object.
}

