% Generated by roxygen2 (4.0.2): do not edit by hand
\name{regex<-.regexr}
\alias{regex<-.regexr}
\title{Set Regex Chunks From a regexr Object}
\usage{
\method{regex}{regexr}(x) <- value
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{value}{The comment(s) to assign.}

\item{\ldots}{Ignored.}
}
\description{
\code{regex<-} - Set the \code{regex} chunk(s) of a \code{regexr} object.
}
\details{
regexr Method for regex<-
}

