% Generated by roxygen2 (4.0.2): do not edit by hand
\name{regex.regexr}
\alias{regex.regexr}
\title{Get Regex Chunks From a regexr Object}
\usage{
\method{regex}{regexr}(x, ...)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\description{
\code{regex} - Get the \code{regex} chunks from a \code{regexr} object.
}
\details{
regexr Method for regex
}

