% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{regexr}
\alias{package-regexr}
\alias{regexr}
\alias{regexr-package}
\title{regexr: Regular Expression Removal, Extraction, & Replacement Tools for the \pkg{qdap} Package}
\description{
\pkg{regexr} is a collection of  regular expression tools associated with
the \pkg{qdap} package that may be useful outside of the context of discourse
analysis.  Tools include removal/extraction/replacement of abbreviations,
dates, dollar amounts, email addresses, hash tags, numbers, percentages,
person tags, phone numbers, times, and zip codes.
}
\details{
The \pkg{regexr} package does not aim to compete with string manipulation
packages such as
\href{http://cran.r-project.org/web/packages/stringr/index.html}{\pkg{stringr}}
or \href{http://cran.r-project.org/web/packages/stringi/index.html}{\pkg{stringi}}
but is meant to provide access to canned, common regular expression patterns
that can be used within \pkg{regexr}, with \pkg{R}'s own regular
expression functions, or add on string manipulation packages such as
\code{stringr} and \code{stringi}.
}

