% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_names}
\alias{set_names}
\title{Set the Names in a \code{regexr} Object}
\usage{
set_names(x, y)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{y}{The names to assign.}
}
\value{
Returns a \code{regexr} object.
}
\description{
This is a convenience function that sets the names on a \code{regexr} object
and returns the object. This function works the same as
\code{\link[stats]{setNames}} but provides a naming which is consistent with
\code{set_regex} and \code{set_comments}.
}
\examples{
minimal <- construct("a", "b", "c")
out <- set_names(minimal, 1:3)
names(out)
}
\seealso{
\code{\link[stats]{setNames}}
}

