% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{subs<-.regexr}
\alias{subs<-.regexr}
\title{Set Regex Sub-expressions From a regexr Object}
\usage{
\method{subs}{regexr}(x) <- value
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{value}{The comment(s) to assign.}

\item{\ldots}{Ignored.}
}
\description{
\code{subs<-} - Set the sub-expressions of a \code{regexr} object.
}
\details{
regexr Method for subs<-
}

