% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subs.regexr}
\alias{subs.regexr}
\title{Get Sub-expressions From a regexr Object}
\usage{
\method{subs}{regexr}(x, ...)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\description{
\code{subs} - Get the sub-expressions from a \code{regexr} object.
}
\details{
regexr Method for subs
}

