% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{summary.regexr}
\alias{summary.regexr}
\title{Summarize a regexr Object}
\usage{
\method{summary}{regexr}(object, ...)
}
\arguments{
\item{object}{The \code{regexr} object}

\item{\ldots}{Ignored.}
}
\description{
Summarize a \code{regexr} object.
}

