% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{test.regexr}
\alias{test.regexr}
\title{Test Regular Expression Validity}
\usage{
\method{test}{regexr}(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{quiet}{logical.  Should \code{test} print warnings about the
concatenated expression and individual sub-expressions?}

\item{\ldots}{Ignored.}
}
\description{
Test regular expression validity of a \code{regexr} object.
}
\details{
test Method for subs<-
}

