% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unglue}
\alias{unglue}
\title{Get Regular Expression Chunks from \code{regexr} Object}
\usage{
unglue(x, ...)
}
\arguments{
\item{x}{A \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\value{
Returns a list of regular expression chunks.
}
\description{
Get regular expression chunks from \code{regexr} object.
}
\examples{
minimal <- construct("a", "b", "c")
minimal
unglue(minimal)
}

