% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{unglue.regexr}
\alias{unglue.regexr}
\title{Unglue regexr Object}
\usage{
\method{unglue}{regexr}(x, ...)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\description{
\code{regexr} - unglue \code{regexr} object.
}
\details{
regexr Method for unglue
}

