% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.regexr.character}
\alias{as.regexr.character}
\title{Coerce character to regexr}
\usage{
\method{as.regexr}{character}(x, ...)
}
\arguments{
\item{x}{The \code{character} object.}

\item{\ldots}{Ignored.}
}
\description{
Convert a regular expression to a commented \code{regexr} object.
}
\details{
character Method for as.regexr
}

