% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.regexr.default}
\alias{as.regexr.default}
\title{Coerce default to regexr}
\usage{
\method{as.regexr}{default}(x, ...)
}
\arguments{
\item{x}{The object to be corced to \code{regexr}.}

\item{\ldots}{Ignored.}
}
\description{
Convert a regular expression to a commented \code{regexr} object.
}
\details{
default Method for as.regexr
}

