% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{comments.regexr}
\alias{comments.regexr}
\title{Extract Comments From regexr Object}
\usage{
\method{comments}{regexr}(x, ...)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\description{
\code{regexr} - Extract comments from \code{regexr} object.
}
\details{
regexr Method for comments
}

