% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regexr-class.R
\name{names.regexr}
\alias{names.regexr}
\title{Get Names of Sub-Expressions of a regexr Object}
\usage{
\method{names}{regexr}(x, ...)
}
\arguments{
\item{x}{The \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\description{
Get names of a \code{regexr} object.
}

