% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{regexr}
\alias{package-regexr}
\alias{regexr}
\alias{regexr-package}
\title{regexr: Tools for Human Readable Regular Expressions}
\description{
\pkg{regexr} is an R framework for constructing human readable regular
expressions.  It aims to provide tools that enable the user to write regular
expressions in a way that is similar to the ways R code is written.
The tools allow the user to:
\enumerate{
  \item Write in smaller, modular, named, \emph{sub-expressions}
  \item Write top to bottom, rather than a single string
  \item Comment individual \emph{sub-expressions}
  \item Indent expressions to represent regular expression groups
  \item Add vertical line spaces and R comments (i.e., \code{#})
  \item Test the validity of the \emph{concatenated expression} and the modular sub-expressions
}
This framework harnesses the power and flexibility of regular expressions
but provides a structural frame that is more consistent with both code
writing and natural language conventions.
}

