% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{regexr}
\alias{package-regexr}
\alias{regexr}
\alias{regexr-package}
\title{regexr: Tools for Human Readable Regular Expressions}
\description{
\pkg{regexr} is an R framework for constructing human readable regular
expressions.  It aims to provide tools that enable the user to write regular
expressions in a way that is similar to the ways R code is written.
The tools allow the user to (1) write in smaller, modular, named, regular
expression chunks, (2) write top to bottom, rather than a single string (3)
comment individual chunks, (4) indent expressions to represent regular
expression groups, and (5) test the validity of the concatenated expression
and the modular chunks. \cr \cr
This framework harnesses the power and flexibility of regular expressions
but provides a structural frame that is more consistent with both coding
writing and natural language conventions.
}

