% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unglue.R
\name{unglue}
\alias{unglue}
\title{Get Sub-expressions from \code{regexr} Object}
\usage{
unglue(x, ...)
}
\arguments{
\item{x}{A \code{regexr} object.}

\item{\ldots}{Ignored.}
}
\value{
Returns a list of regular expression chunks.
}
\description{
Get sub-expressions from \code{regexr} object.
}
\examples{
minimal <- construct("a", "b", "c")
minimal
unglue(minimal)
}

